// src/modules/balancer-sdk/sor/impl/types.ts
export enum SwapTypes {
  SwapExactIn,
  SwapExactOut,
}

export enum PoolTypes {
  Weighted,
  Stable,
  Element,
  MetaStable,
  Linear,
  Gyro2,
  Gyro3,
  GyroE,
}

export interface SwapOptions {
  gasPrice: BigNumber;
  swapGas: BigNumber;
  timestamp: number;
  maxPools: number;
  poolTypeFilter: PoolFilter;
  forceRefresh: boolean;
}

export type PoolPairBase = {
  id: string;
  address: string;
  poolType: PoolTypes;
  swapFee: BigNumber;
  tokenIn: string;
  tokenOut: string;
  decimalsIn: number;
  decimalsOut: number;
  balanceIn: BigNumber;
  balanceOut: BigNumber;
};

export interface Swap {
  pool: string;
  tokenIn: string;
  tokenOut: string;
  swapAmount?: string;
  limitReturnAmount?: string;
  maxPrice?: string;
  tokenInDecimals: number;
  tokenOutDecimals: number;
  returnAmount?: string;
}

export interface SubgraphPoolBase {
  id: string;
  address: string;
  poolType: string;
  swapFee: string;
  swapEnabled: boolean;
  totalShares: string;
  tokens: SubgraphToken[];
  tokensList: string[];

  // Weighted & Element field
  totalWeight?: string;

  // Stable specific fields
  amp?: string;

  // Element specific fields
  expiryTime?: number;
  unitSeconds?: number;
  principalToken?: string;
  baseToken?: string;

  // Linear specific fields
  mainIndex?: number;
  wrappedIndex?: number;
  lowerTarget?: string;
  upperTarget?: string;

  // Gyro2 specific fields
  sqrtAlpha?: string;
  sqrtBeta?: string;

  // Gyro3 specific field
  root3Alpha?: string;

  // GyroE specific fields
  alpha?: string;
  beta?: string;
  c?: string;
  s?: string;
  lambda?: string;
  tauAlphaX?: string;
  tauAlphaY?: string;
  tauBetaX?: string;
  tauBetaY?: string;
  u?: string;
  v?: string;
  w?: string;
  z?: string;
  dSq?: string;
}

export type SubgraphToken = {
  address: string;
  balance: string;
  decimals: number;
  priceRate: string;
  // WeightedPool field
  weight: string | null;
};
// ...
export interface PoolBase {
  poolType: PoolTypes;
  id: string;
  address: string;
  tokensList: string[];
  tokens: { address: string; balance: string; decimals: number }[];
  totalShares: BigNumber;
  mainIndex?: number;
  isLBP?: boolean;
  parsePoolPairData: (tokenIn: string, tokenOut: string) => PoolPairBase;
  getNormalizedLiquidity: (poolPairData: PoolPairBase) => OldBigNumber;
  getLimitAmountSwap: (poolPairData: PoolPairBase, swapType: SwapTypes) => OldBigNumber;
  /**
   * @param {string} token - Address of token.
   * @param {BigNumber} newBalance - New balance of token. EVM scaled.
   */
  updateTokenBalanceForPool: (token: string, newBalance: BigNumber) => void;
  _exactTokenInForTokenOut: (poolPairData: PoolPairBase, amount: OldBigNumber) => OldBigNumber;
  _tokenInForExactTokenOut: (poolPairData: PoolPairBase, amount: OldBigNumber) => OldBigNumber;
  _calcTokensOutGivenExactBptIn(bptAmountIn: BigNumber): BigNumber[];
  _calcBptOutGivenExactTokensIn(amountsIn: BigNumber[]): BigNumber;
  _spotPriceAfterSwapExactTokenInForTokenOut: (
    poolPairData: PoolPairBase,
    amount: OldBigNumber,
  ) => OldBigNumber;
  _spotPriceAfterSwapTokenInForExactTokenOut: (
    poolPairData: PoolPairBase,
    amount: OldBigNumber,
  ) => OldBigNumber;
  _derivativeSpotPriceAfterSwapExactTokenInForTokenOut: (
    poolPairData: PoolPairBase,
    amount: OldBigNumber,
  ) => OldBigNumber;
  _derivativeSpotPriceAfterSwapTokenInForExactTokenOut: (
    poolPairData: PoolPairBase,
    amount: OldBigNumber,
  ) => OldBigNumber;
}